made helper funcs
