+++ controllers/userResumeController.js
@@ -2,8 +2,9 @@ var express = require('express')
 var mongoose = require('mongoose');
 var router = express.Router();
 var ResumeData = require('../models/resume_data.js');
-var UserResume = require('../models/user_resume.js')
-var User = require('../models/user.js')
+var UserResume = require('../models/user_resume.js');
+var User = require('../models/user.js');
+var helper = require('../helpers/resume_form.js')
 
 
 
@@ -36,45 +37,8 @@ module.exports.controller = function(app) {
     //sub == 'submission'
     var sub = req.body
       //start new function
-    var newResume = new UserResume();
-    //make a new resume, give all submission data to its appropriate keys
-    newResume.name = sub.Name;
-    newResume.streetAddress = sub.Address;
-    newResume.email = sub.Email;
-    newResume.phone = sub.Phone;
-    newResume.education = sub.Education
-    newResume.summary = sub.Summary
-      //create a new experience object for the array of experience for each new experience that gets added
-    newResume.experiences = [];
-    var experienceObject = {}
-    sub.textinput.forEach(function(input, index) {
-        switch (index % 4) {
-          case 0:
-            experienceObject.title = input
-            break;
-          case 1:
-            experienceObject.startDate = input;
-            break;
-          case 2:
-            experienceObject.endDate = input;
-            break;
-          case 3:
-            experienceObject.description = input;
-            newResume.experiences.push(experienceObject)
-            break;
-        }
-      })
-      //give new resume to the current user
-    req.user.resumes.push(newResume);
-    //save unpdated current user and the resume itself into the db
-    newResume.save(function(err) {
-      if (err)
-        throw err;
-    });
-    req.user.save(function(err) {
-      if (err)
-        throw err;
-    }); //end new function
+      helper.newResume(sub, req);
+     //end new function
     res.render('profile', {
       req: req,
       user: req.user
@@ -89,70 +53,13 @@ module.exports.controller = function(app) {
         //delete the old resume (from user and database)
       var resumeID = sub.ID;
       //delete function:
-      UserResume.remove({
-          _id: resumeID
-        }, function(err, result) {
-          if (err) {
-            console.log(err);
-          } else {
-            console.log('deleted');
-            var userID = req.user._id;
-            User.findById(userID, function(err, user) {
-              if (err) throw err;
-              user.resumes.forEach(function(resume, index) {
-                if (resume._id == resumeID) {
-                  user.resumes.splice(index, 1);
-                  return
-                }
-              })
-              user.save(function(err) {
-                if (err) throw err;
-              })
-            })
-          }
-        }) //end delete function
+      helper.deleteResume(resumeID, req)
+       //end delete function
 
       var sub = req.body
         //start new function
-      var newResume = new UserResume();
-      //make a new resume, give all submission data to its appropriate keys
-      newResume.name = sub.Name;
-      newResume.streetAddress = sub.Address;
-      newResume.email = sub.Email;
-      newResume.phone = sub.Phone;
-      newResume.education = sub.Education
-      newResume.summary = sub.Summary
-        //create a new experience object for the array of experience for each new experience that gets added
-      newResume.experiences = [];
-      var experienceObject = {}
-      sub.textinput.forEach(function(input, index) {
-          switch (index % 4) {
-            case 0:
-              experienceObject.title = input
-              break;
-            case 1:
-              experienceObject.startDate = input;
-              break;
-            case 2:
-              experienceObject.endDate = input;
-              break;
-            case 3:
-              experienceObject.description = input;
-              newResume.experiences.push(experienceObject)
-              break;
-          }
-        })
-        //give new resume to the current user
-      req.user.resumes.push(newResume);
-      //save unpdated current user and the resume itself into the db
-      newResume.save(function(err) {
-        if (err)
-          throw err;
-      });
-      req.user.save(function(err) {
-        if (err)
-          throw err;
-      }); //end new function
+            helper.newResume(sub, req);
+      //end new function
 
       res.redirect('/profile')
 
@@ -164,28 +71,8 @@ module.exports.controller = function(app) {
   app.get('/resume/delete/:id', function(req, res) {
     var resumeID = req.params.id
       //delete function:
-    UserResume.remove({
-        _id: resumeID
-      }, function(err, result) {
-        if (err) {
-          console.log(err);
-        } else {
-          console.log('deleted');
-          var userID = req.user._id;
-          User.findById(userID, function(err, user) {
-            if (err) throw err;
-            user.resumes.forEach(function(resume, index) {
-              if (resume._id == resumeID) {
-                user.resumes.splice(index, 1);
-                return
-              }
-            })
-            user.save(function(err) {
-              if (err) throw err;
-            })
-          })
-        }
-      }) //end delete function
+        helper.deleteResume(resumeID, req)
+   //end delete function
     res.redirect('/profile')
   })
